refactor: Split persistence.rs into modular save/load structure
Split the 615-line persistence.rs file into a focused module structure:

- `persistence/mod.rs`: Module organization and documentation
- `persistence/save.rs`: SQL dump generation logic (~260 lines)
- `persistence/tests.rs`: All persistence tests (~355 lines)

Benefits:
- Separation of concerns: Reading and writing operations isolated
- Easier testing: Each operation can be tested independently
- Reduced file size: Each file now ~300 lines instead of 600+
- Ready for future load operations: Structure prepared for load.rs

All existing tests pass with no functional changes to the API.
Public API remains unchanged (backwards compatible).

Closes #1026
