--- crates/storage/src/persistence.rs
@@ -1,615 +1,0 @@
-// ============================================================================
-// Database Persistence - SQL Dump Format
-// ============================================================================
-//
-// Provides SQL dump format (human-readable, portable) for database persistence.
-// Generates SQL statements that recreate the database state including schemas,
-// tables, data, indexes, roles, and privileges.
-
-use crate::{Database, StorageError};
-use std::fs::File;
-use std::io::{BufWriter, Write};
-use std::path::Path;
-
-impl Database {
-    /// Save database state as SQL dump (human-readable, portable)
-    ///
-    /// Generates SQL statements that recreate the database state including:
-    /// - Schemas
-    /// - Tables with column definitions
-    /// - Indexes
-    /// - Data (INSERT statements)
-    /// - Roles and privileges
-    ///
-    /// # Example
-    /// ```no_run
-    /// # use storage::Database;
-    /// let db = Database::new();
-    /// db.save_sql_dump("database.sql").unwrap();
-    /// ```
-    pub fn save_sql_dump<P: AsRef<Path>>(&self, path: P) -> Result<(), StorageError> {
-        let file = File::create(path)
-            .map_err(|e| StorageError::NotImplemented(format!("Failed to create file: {}", e)))?;
-
-        let mut writer = BufWriter::new(file);
-
-        // Header
-        writeln!(writer, "-- VibeSQL Database Dump")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        writeln!(writer, "-- Generated: {}", chrono::Utc::now())
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        writeln!(writer, "--")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        writeln!(writer)
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        // Export schemas (except default 'public' which always exists)
-        writeln!(writer, "-- Schemas")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        for schema_name in self.catalog.list_schemas() {
-            if schema_name != "public" {
-                writeln!(writer, "CREATE SCHEMA {};", schema_name)
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-            }
-        }
-        writeln!(writer)
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        // Export roles
-        writeln!(writer, "-- Roles")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        for role_name in self.catalog.list_roles() {
-            writeln!(writer, "CREATE ROLE {};", role_name)
-                .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        }
-        writeln!(writer)
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        // Export tables and data
-        writeln!(writer, "-- Tables and Data")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        // Get list of table names
-        let table_names = self.catalog.list_tables();
-
-        for table_name in &table_names {
-            if let Some(table) = self.get_table(&table_name) {
-                // CREATE TABLE statement
-                let schema = &table.schema;
-                write!(writer, "CREATE TABLE {} (", &table_name)
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-                for (i, col) in schema.columns.iter().enumerate() {
-                    if i > 0 {
-                        write!(writer, ", ")
-                            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                    }
-                    write!(writer, "{} {}", col.name, format_data_type(&col.data_type))
-                        .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                    if !col.nullable {
-                        write!(writer, " NOT NULL")
-                            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                    }
-                }
-
-                writeln!(writer, ");")
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-                // INSERT statements for data
-                if table.row_count() > 0 {
-                    writeln!(writer)
-                        .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                    for row in table.scan() {
-                        write!(writer, "INSERT INTO {} VALUES (", &table_name)
-                            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                        for (i, value) in row.values.iter().enumerate() {
-                            if i > 0 {
-                                write!(writer, ", ")
-                                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                            }
-                            write!(writer, "{}", sql_value_to_literal(value))
-                                .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                        }
-                        writeln!(writer, ");")
-                            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                    }
-                }
-
-                writeln!(writer)
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-            }
-        }
-
-        // Export indexes
-        writeln!(writer, "-- Indexes")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        for index_name in self.list_indexes() {
-            let metadata = self.get_index(&index_name).unwrap();
-            write!(writer, "CREATE")
-                .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-            if metadata.unique {
-                write!(writer, " UNIQUE")
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-            }
-            write!(writer, " INDEX {} ON {} (", index_name, metadata.table_name)
-                .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-            for (i, col) in metadata.columns.iter().enumerate() {
-                if i > 0 {
-                    write!(writer, ", ")
-                        .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                }
-                write!(writer, "{}", col.column_name)
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-                write!(writer, " {:?}", col.direction)
-                    .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-            }
-
-            writeln!(writer, ");")
-                .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-        }
-        writeln!(writer)
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        writeln!(writer, "-- End of dump")
-            .map_err(|e| StorageError::NotImplemented(format!("Write error: {}", e)))?;
-
-        Ok(())
-    }
-}
-
-/// Format a DataType for SQL CREATE TABLE statement
-fn format_data_type(data_type: &types::DataType) -> String {
-    use types::DataType;
-
-    match data_type {
-        DataType::Integer => "INTEGER".to_string(),
-        DataType::Smallint => "SMALLINT".to_string(),
-        DataType::Bigint => "BIGINT".to_string(),
-        DataType::Unsigned => "BIGINT UNSIGNED".to_string(),
-        DataType::Float { .. } => "FLOAT".to_string(),
-        DataType::Real => "REAL".to_string(),
-        DataType::DoublePrecision => "DOUBLE PRECISION".to_string(),
-        DataType::Varchar { max_length } => {
-            if let Some(len) = max_length {
-                format!("VARCHAR({})", len)
-            } else {
-                "VARCHAR".to_string()
-            }
-        }
-        DataType::Character { length } => format!("CHAR({})", length),
-        DataType::Boolean => "BOOLEAN".to_string(),
-        DataType::Date => "DATE".to_string(),
-        DataType::Time { .. } => "TIME".to_string(),
-        DataType::Timestamp { .. } => "TIMESTAMP".to_string(),
-        DataType::Interval { .. } => "INTERVAL".to_string(),
-        DataType::Numeric { precision, scale } => {
-            format!("NUMERIC({}, {})", precision, scale)
-        }
-        DataType::Timestamp { with_timezone } => {
-            if *with_timezone {
-                "TIMESTAMP WITH TIME ZONE".to_string()
-            } else {
-                "TIMESTAMP".to_string()
-            }
-        }
-        DataType::Interval { start_field, end_field: _ } => {
-            // Simplified interval representation for now
-            format!("INTERVAL {:?}", start_field)
-        }
-        DataType::Numeric { precision, scale } => {
-            format!("NUMERIC({}, {})", precision, scale)
-        }
-        DataType::Decimal { precision, scale } => {
-            format!("DECIMAL({}, {})", precision, scale)
-        }
-        DataType::CharacterLargeObject => "CLOB".to_string(),
-        DataType::Name => "VARCHAR(128)".to_string(),
-        DataType::BinaryLargeObject => "BLOB".to_string(),
-        DataType::UserDefined { type_name } => type_name.clone(),
-        DataType::Null => "NULL".to_string(),
-    }
-}
-
-/// Convert a SqlValue to its SQL literal representation
-fn sql_value_to_literal(value: &types::SqlValue) -> String {
-    use types::SqlValue;
-
-    match value {
-        SqlValue::Null => "NULL".to_string(),
-        SqlValue::Integer(n) => n.to_string(),
-        SqlValue::Smallint(n) => n.to_string(),
-        SqlValue::Bigint(n) => n.to_string(),
-        SqlValue::Unsigned(n) => n.to_string(),
-        SqlValue::Numeric(f) => f.to_string(),
-        SqlValue::Float(f) | SqlValue::Real(f) => {
-            if f.is_nan() {
-                "'NaN'".to_string()
-            } else if f.is_infinite() {
-                if f.is_sign_positive() {
-                    "'Infinity'".to_string()
-                } else {
-                    "'-Infinity'".to_string()
-                }
-            } else {
-                f.to_string()
-            }
-        }
-        SqlValue::Double(f) => {
-            if f.is_nan() {
-                "'NaN'".to_string()
-            } else if f.is_infinite() {
-                if f.is_sign_positive() {
-                    "'Infinity'".to_string()
-                } else {
-                    "'-Infinity'".to_string()
-                }
-            } else {
-                f.to_string()
-            }
-        }
-        SqlValue::Character(s) | SqlValue::Varchar(s) => format!("'{}'", s.replace('\'', "''")),
-        SqlValue::Boolean(b) => if *b { "TRUE" } else { "FALSE" }.to_string(),
-        SqlValue::Date(d) => format!("DATE '{}'", d),
-        SqlValue::Time(t) => format!("TIME '{}'", t),
-        SqlValue::Timestamp(ts) => format!("TIMESTAMP '{}'", ts),
-        SqlValue::Interval(i) => format!("INTERVAL '{}'", i),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use catalog::{ColumnSchema, TableSchema};
-    use types::{DataType, SqlValue};
-
-    #[test]
-    fn test_save_sql_dump() {
-        let mut db = Database::new();
-
-        // Create test schema
-        let schema = TableSchema::new(
-            "test".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("name".to_string(), DataType::Varchar { max_length: Some(50) }, false),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Insert test data
-        let table = db.get_table_mut("test").unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(1), SqlValue::Varchar("Alice".to_string())])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(2), SqlValue::Varchar("Bob".to_string())])).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_db.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify file exists and contains expected content
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("CREATE TABLE"));
-        assert!(content.contains("INSERT INTO"));
-        assert!(content.contains("Alice"));
-        assert!(content.contains("Bob"));
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_value_to_literal() {
-        assert_eq!(sql_value_to_literal(&SqlValue::Null), "NULL");
-        assert_eq!(sql_value_to_literal(&SqlValue::Integer(42)), "42");
-        assert_eq!(sql_value_to_literal(&SqlValue::Varchar("test".to_string())), "'test'");
-        assert_eq!(sql_value_to_literal(&SqlValue::Varchar("test's".to_string())), "'test''s'");
-        assert_eq!(sql_value_to_literal(&SqlValue::Boolean(true)), "TRUE");
-        assert_eq!(sql_value_to_literal(&SqlValue::Boolean(false)), "FALSE");
-    }
-
-    #[test]
-    fn test_sql_dump_with_nulls() {
-        let mut db = Database::new();
-
-        // Create test schema with nullable column
-        let schema = TableSchema::new(
-            "test_nulls".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("nullable_col".to_string(), DataType::Varchar { max_length: Some(50) }, true),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Insert test data with NULL values
-        let table = db.get_table_mut("test_nulls").unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(1), SqlValue::Null])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(2), SqlValue::Varchar("text".to_string())])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(3), SqlValue::Null])).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_nulls.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify SQL contains NULL literal
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("NULL"), "SQL dump should contain NULL literal");
-        assert!(content.contains("INSERT INTO test_nulls VALUES (1, NULL)"));
-        assert!(content.contains("INSERT INTO test_nulls VALUES (2, 'text')"));
-        assert!(content.contains("INSERT INTO test_nulls VALUES (3, NULL)"));
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_with_quotes() {
-        let mut db = Database::new();
-
-        // Create test schema
-        let schema = TableSchema::new(
-            "test_quotes".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("text".to_string(), DataType::Varchar { max_length: Some(100) }, false),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Insert test data with various quote types
-        let table = db.get_table_mut("test_quotes").unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(1), SqlValue::Varchar("it's".to_string())])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(2), SqlValue::Varchar("say \"hello\"".to_string())])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(3), SqlValue::Varchar("it's a \"test\"".to_string())])).unwrap();
-        table.insert(crate::Row::new(vec![SqlValue::Integer(4), SqlValue::Varchar("".to_string())])).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_quotes.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify SQL properly escapes quotes
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("'it''s'"), "Single quotes should be escaped as ''");
-        assert!(content.contains("'say \"hello\"'"), "Double quotes should be preserved in single-quoted strings");
-        assert!(content.contains("'it''s a \"test\"'"), "Mixed quotes should be handled correctly");
-        assert!(content.contains("''"), "Empty string should be represented");
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_empty_database() {
-        let db = Database::new();
-
-        // Save SQL dump of empty database
-        let path = "/tmp/test_empty_db.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify file exists and contains valid SQL structure
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("-- VibeSQL Database Dump"));
-        assert!(content.contains("-- Schemas"));
-        assert!(content.contains("-- Roles"));
-        assert!(content.contains("-- Tables and Data"));
-        assert!(content.contains("-- Indexes"));
-        assert!(content.contains("-- End of dump"));
-
-        // Verify no CREATE TABLE statements in empty database
-        assert!(!content.contains("CREATE TABLE"));
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_empty_table() {
-        let mut db = Database::new();
-
-        // Create test schema but insert no data
-        let schema = TableSchema::new(
-            "empty_table".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("name".to_string(), DataType::Varchar { max_length: Some(50) }, false),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_empty_table.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify SQL contains CREATE TABLE but no INSERT statements
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("CREATE TABLE empty_table"));
-        assert!(!content.contains("INSERT INTO empty_table"), "Empty table should have no INSERT statements");
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_with_indexes() {
-        let mut db = Database::new();
-
-        // Create test schema
-        let schema = TableSchema::new(
-            "test_indexes".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("name".to_string(), DataType::Varchar { max_length: Some(50) }, false),
-                ColumnSchema::new("email".to_string(), DataType::Varchar { max_length: Some(100) }, false),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Create indexes (use fully qualified table name)
-        let idx1 = ast::IndexColumn {
-            column_name: "name".to_string(),
-            direction: ast::OrderDirection::Asc,
-        };
-        db.create_index("idx_name".to_string(), "public.test_indexes".to_string(), false, vec![idx1]).unwrap();
-
-        let idx2 = ast::IndexColumn {
-            column_name: "email".to_string(),
-            direction: ast::OrderDirection::Asc,
-        };
-        db.create_index("idx_email_unique".to_string(), "public.test_indexes".to_string(), true, vec![idx2]).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_indexes.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify SQL contains CREATE INDEX statements
-        let content = std::fs::read_to_string(path).unwrap();
-        // Index names are normalized to uppercase
-        assert!(content.contains("CREATE INDEX IDX_NAME"), "Should contain CREATE INDEX IDX_NAME");
-        assert!(content.contains("CREATE UNIQUE INDEX IDX_EMAIL_UNIQUE"), "Should contain CREATE UNIQUE INDEX IDX_EMAIL_UNIQUE");
-        assert!(content.contains("ON public.test_indexes"), "Should reference test_indexes table");
-        assert!(content.contains("Asc"), "Index direction should be included");
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_all_data_types() {
-        let mut db = Database::new();
-
-        // Create test schema with all data types
-        let schema = TableSchema::new(
-            "all_types".to_string(),
-            vec![
-                // Integer types
-                ColumnSchema::new("col_int".to_string(), DataType::Integer, true),
-                ColumnSchema::new("col_smallint".to_string(), DataType::Smallint, true),
-                ColumnSchema::new("col_bigint".to_string(), DataType::Bigint, true),
-                // Float types
-                ColumnSchema::new("col_float".to_string(), DataType::Float { precision: 24 }, true),
-                ColumnSchema::new("col_real".to_string(), DataType::Real, true),
-                ColumnSchema::new("col_double".to_string(), DataType::DoublePrecision, true),
-                // String types
-                ColumnSchema::new("col_varchar".to_string(), DataType::Varchar { max_length: Some(100) }, true),
-                ColumnSchema::new("col_char".to_string(), DataType::Character { length: 10 }, true),
-                // Boolean
-                ColumnSchema::new("col_bool".to_string(), DataType::Boolean, true),
-                // Numeric
-                ColumnSchema::new("col_numeric".to_string(), DataType::Numeric { precision: 10, scale: 2 }, true),
-                ColumnSchema::new("col_decimal".to_string(), DataType::Decimal { precision: 5, scale: 0 }, true),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Insert test data with various values
-        let table = db.get_table_mut("all_types").unwrap();
-        table.insert(crate::Row::new(vec![
-            SqlValue::Integer(42),
-            SqlValue::Smallint(100),
-            SqlValue::Bigint(999999),
-            SqlValue::Float(3.14),
-            SqlValue::Real(2.718),
-            SqlValue::Double(1.414),
-            SqlValue::Varchar("test".to_string()),
-            SqlValue::Character("fixed".to_string()),
-            SqlValue::Boolean(true),
-            SqlValue::Numeric(123.45),
-            SqlValue::Numeric(999.0),
-        ])).unwrap();
-
-        // Test special float values (NaN, Infinity)
-        table.insert(crate::Row::new(vec![
-            SqlValue::Null,
-            SqlValue::Null,
-            SqlValue::Null,
-            SqlValue::Float(f32::NAN),
-            SqlValue::Real(f32::INFINITY),
-            SqlValue::Double(f64::NEG_INFINITY),
-            SqlValue::Null,
-            SqlValue::Null,
-            SqlValue::Boolean(false),
-            SqlValue::Null,
-            SqlValue::Null,
-        ])).unwrap();
-
-        // Save SQL dump
-        let path = "/tmp/test_all_types.sql";
-        db.save_sql_dump(path).unwrap();
-
-        // Verify SQL contains all data type declarations
-        let content = std::fs::read_to_string(path).unwrap();
-        assert!(content.contains("INTEGER"));
-        assert!(content.contains("SMALLINT"));
-        assert!(content.contains("BIGINT"));
-        assert!(content.contains("FLOAT(24)"));
-        assert!(content.contains("REAL"));
-        assert!(content.contains("DOUBLE PRECISION"));
-        assert!(content.contains("VARCHAR(100)"));
-        assert!(content.contains("CHAR(10)"));
-        assert!(content.contains("BOOLEAN"));
-        assert!(content.contains("NUMERIC(10, 2)"));
-        assert!(content.contains("DECIMAL(5, 0)"));
-
-        // Verify special float values
-        assert!(content.contains("'NaN'"));
-        assert!(content.contains("'Infinity'"));
-        assert!(content.contains("'-Infinity'"));
-        assert!(content.contains("TRUE"));
-        assert!(content.contains("FALSE"));
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-
-    #[test]
-    fn test_sql_dump_large_dataset() {
-        let mut db = Database::new();
-
-        // Create test schema
-        let schema = TableSchema::new(
-            "large_table".to_string(),
-            vec![
-                ColumnSchema::new("id".to_string(), DataType::Integer, false),
-                ColumnSchema::new("value".to_string(), DataType::Varchar { max_length: Some(50) }, false),
-            ],
-        );
-
-        db.create_table(schema).unwrap();
-
-        // Insert 10,000 rows
-        let table = db.get_table_mut("large_table").unwrap();
-        for i in 0..10000 {
-            table.insert(crate::Row::new(vec![
-                SqlValue::Integer(i as i64),
-                SqlValue::Varchar(format!("value_{}", i)),
-            ])).unwrap();
-        }
-
-        // Save SQL dump and measure
-        let path = "/tmp/test_large.sql";
-        let start = std::time::Instant::now();
-        db.save_sql_dump(path).unwrap();
-        let duration = start.elapsed();
-
-        // Verify export completed in reasonable time (should be fast)
-        assert!(duration.as_secs() < 10, "Export took too long: {:?}", duration);
-
-        // Verify file exists and has content
-        let metadata = std::fs::metadata(path).unwrap();
-        assert!(metadata.len() > 0, "File should have content");
-
-        // Verify row count in file
-        let content = std::fs::read_to_string(path).unwrap();
-        let insert_count = content.matches("INSERT INTO large_table").count();
-        assert_eq!(insert_count, 10000, "Should have 10,000 INSERT statements");
-
-        // Cleanup
-        std::fs::remove_file(path).ok();
-    }
-}
